ai-multi-agent-system/requirements.txt
# Core Dependencies
crewai>=0.28.0
crewai-tools>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Azure AI SDK
azure-identity>=1.15.0
openai>=1.12.0

# Communication Channels
# Email
imapclient>=2.3.1
pyzmail36>=1.0.5
smtplib-wrapper>=1.0.0

# WhatsApp
requests>=2.31.0
python-whatsapp-bot>=0.1.0

# Teams
pymsteams>=0.2.2
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP & API
httpx>=0.26.0
aiohttp>=3.9.0

# Logging & Monitoring
loguru>=0.7.2
python-json-logger>=2.0.7

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Database (Optional)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Caching (Optional)
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development Tools
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
